feat(indexer): emit and resolve mod declaration edges
Phase 1a of #121: capture `pub mod xxx;` / `mod xxx;` as edges in
`file_references` so brief's forward closure can include sibling
modules under lib.rs / parent.rs files.

- chunker/rust.rs: parse body-less mod_item as ParsedImport(ModDecl)
- rust_resolver.rs: resolve_mod_decl uses self::name probing; tests
  split to rust_resolver/tests.rs (482 -> 164 lines)
- resolver.rs: Resolve::resolve_mod_decl trait method (default None)
- indexer.rs: build_references factored into resolve_target +
  import_to_references; ModDecl routes via resolve_mod_decl
- storage/types.rs: RefKind::ModDecl ("mod_decl")

Measured on yomu repo (35 files):
- TOTAL_EDGES 118 -> 149 (+31), mod_decl edges 0 -> 30
- src/lib.rs 0 -> 8, src/storage.rs 0 -> 7
- ZERO-EDGE 8 -> 6 (5 leaf modules + main.rs for Phase 1b)

438 lib + 2 integration tests pass (5 new: T-573/T-574, T-334..T-336).
